refactor: reform the register paths
