+++ packages/frontend/app/login/page.tsx
@@ -3,7 +3,7 @@
 import { useState } from "react";
 
 import { LoginForm } from "./components/login-form";
-import { RegisterForm } from "./components/register-form";
+import { RegisterForm } from "../register/user/components/form";
 
 export default function Login() {
   const [formState, setFormState] = useState("login");
