+++ packages/frontend/app/register/user/page.tsx
@@ -1,0 +1,5 @@
+import { RegisterForm } from "./components/form";
+
+export default function Page() {
+  return <RegisterForm />;
+}
