+++ packages/frontend/app/register/weekly-meetings/components/map-view.tsx
