+++ packages/frontend/app/register/weekly-meetings/components/meeting-card.tsx
