+++ packages/frontend/app/register/weekly-meetings/components/weekly-meetings-list.tsx
